Build: make build error fix
This patch is for fixing build error like below.

sometimes the build error occur which
obj/startup.d cannot be found.
The cause is that the obj folder cannot be found by
compiling without performing mkdir first in
the multi-thread environment like -j16.
so the code was modified to execute mkdir first.

Signed-off-by: chungho.shin <chungho.shin@nexell.co.kr>
Change-Id: I3e6848fcf2457099e13c6f8fbd0cc258dbc69d56
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/bl1_s5p6818/+/488940
Reviewed-by: allan park <allan.park@nexell.co.kr>
Tested-by: allan park <allan.park@nexell.co.kr>